modificacion gestion elo en bdd
Con esto queda zanjado por ahora como gestionar los 4 elos. Doy por terminadas las tareas de diseño y creacion de tablas en bdd. closes #5 closes #2

+++ internal/db/schema.sql
@@ -116,7 +116,7 @@ CREATE TABLE IF NOT EXISTS "friendship" (
 CREATE TABLE IF NOT EXISTS "rating" (
 	"user_id" BIGINT NOT NULL,
 	"elo_type" ELO_TYPE NOT NULL,
-	"value" BIGINT NOT NULL,
+	"value" INT NOT NULL,
 	PRIMARY KEY("user_id", "elo_type"),
 	FOREIGN KEY("user_id") REFERENCES "account"("account_id")
 );
